kotlin/geography: Stop relying on LocationManager#getCurrentLocation()
Although the method includes an executor as one of its parameters, it is
only effective for the callback. This deviates from the expectations set
by the subsequent flows, where entities need to fulfill their contracts
synchronously.
